implement E.... E...... feedback
Your code is simple and clear, but your tests don't have enough coverage.
In Rover tests, you use the production code itself for the expectations. You're checking that when an 'F' command is sent to the Rover, its location should be equal to the outcome of the "Location.forward" method. And this could be just fine if you ensure that the "Location.forward" method is doing the right thing. And the Location tests are not assuring that. They are not testing that only the x value changes when the Rover receives an 'F' command, for instance. I could change the "Location.forward" method to something like this:

public void forward() {
    switch (direction) {
        case <em>EAST </em>-> {
            x += 1;
            y += 1;
        }
        case <em>NORTH </em>-> y += 1;
        case <em>WEST </em>-> x -= 1;
        case <em>SOUTH </em>-> y -= 1;
    }
}

and all the tests still pass