implement S..... S....... feedback
value value in enum has a specific name, ordinal. Direction.<NORTH/EAST/SOUTH/WEST>.ordinal() will give the same value. It's dangerous in a sense, that enum can be altered and ordinals would change. But in case of "directions", it's a safe bet, no one will invent additional direction, as there is none, these are constant values. The risk is if someone would decide to alter enum and add NORTHEAST. Just keep this in mind.